Insets cleanup patch from Angus
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@1127 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/combox.h
@@ -75,7 +75,7 @@ public:
 	string const getline() const;
    
 	///  Select an arbitrary item
-	void select(size_t);
+	void select(int);
 	///
         bool select_text(string const &);
